# MentorLink/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MentorLink API",
        "description": "API RESTful MentorLink — Auth via Bearer token (Sanctum)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/admin/stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "[Admin] Statistiques globales",
                "operationId": "bb126a135b7d8ece3e86bc758930c615",
                "responses": {
                    "200": {
                        "description": "Statistiques"
                    },
                    "403": {
                        "description": "Réservé à l'admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/pending-mentors": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "[Admin] Profils en attente de validation",
                "operationId": "ffc9ead671bde3c7cb5aaf7fd3d937f5",
                "responses": {
                    "200": {
                        "description": "Profils en attente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "mentor",
                                            "mentee"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token retourné"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Déconnecté"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Utilisateur connecté",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/mentors": {
            "get": {
                "tags": [
                    "Mentors"
                ],
                "summary": "Liste des mentors validés",
                "operationId": "e20cd905f0437e91f6850b57aa49cb2a",
                "parameters": [
                    {
                        "name": "domain",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée"
                    }
                }
            }
        },
        "/api/mentors/{id}": {
            "get": {
                "tags": [
                    "Mentors"
                ],
                "summary": "Détail d'un mentor",
                "operationId": "f19ef0f5df60e1702a927f115e7db6be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil mentor"
                    },
                    "404": {
                        "description": "Introuvable"
                    }
                }
            }
        },
        "/api/mentor/profile": {
            "post": {
                "tags": [
                    "Mentors"
                ],
                "summary": "Créer/mettre à jour son profil mentor",
                "operationId": "f770ed157aaa9e16e1381604e33a74b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "domains"
                                ],
                                "properties": {
                                    "domains": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "hourly_rate": {
                                        "type": "number"
                                    },
                                    "bio": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/mentors/{id}/validate": {
            "put": {
                "tags": [
                    "Mentors",
                    "Admin"
                ],
                "summary": "[Admin] Valider un profil mentor",
                "operationId": "314bfeff288741194fda49ff54cd9f2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil validé"
                    },
                    "403": {
                        "description": "Réservé à l'admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}